+++ crates/prelude/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.5.1-git
+## 0.6.0-git
+
+### Major
+
+- Change `button::Listener::new()` to never panic and return an error instead
 
 ### Minor